backend/app/db.py
import os
from typing import Annotated

from dotenv import load_dotenv
from fastapi import Depends
from sqlmodel import Session, SQLModel, create_engine

import app.models

load_dotenv()
engine = create_engine(os.environ.get("DB_URL"), echo=bool(os.environ.get("DEBUG", False)))


def create_db_and_tables():
    SQLModel.metadata.create_all(engine)


def get_session():
    with Session(engine) as session:
        yield session


SessionDep = Annotated[Session, Depends(get_session)]
