backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.118.3",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.27",
]
